This is the start of the Pi sensor node repository.
Current sensors supported include the BMP180 and the start of the DS18B20
thermal sensor.

netlogging is supported, as is local logging.  To start the netlogger, change
the lines in test.py to enable log.netLogger(), and disable log.logPrint.

To start the python sensor, use LocalTP.py.  This will poll the BMP180
every minute, and log back to the default IP address.